Core_Backend_Services/LLM_Morphing_Service/requirements.txt
langgraph 
langchain-core 
langchain-groq 
langchain-google-genai 
pydantic 
sentence-transformers 
langgraph
langchain-core
langchain-groq
langchain-google-genai
pydantic
sentence-transformers
python-dotenv
fastapi
uvicorn
langsmith
pytest
pytest-asyncio
numpy
scikit-learn
sqlalchemy
psycopg2-binary